+++ LegendWorld/Data/Items/ContainerItem.cs
@@ -11,6 +11,9 @@ namespace LegendWorld.Data.Items
         }
 
         public List<ushort> Items { get; set; }
+        //public ushort Id { get; set; }
+        //public ItemIdentity Identity { get; set; }
+        //public ItemCategory Category { get; set; }
 
         public void AddItem(ushort id)
         {
@@ -28,11 +31,11 @@ namespace LegendWorld.Data.Items
         }
         public void AddItem(Item itemToAdd)
         {
-            this.AddItem(itemToAdd.Id);
+            this.AddItem((ushort)itemToAdd.Id);
         }
         public void RemoveItem(Item itemToRemove)
         {
-            this.RemoveItem(itemToRemove.Id);
+            this.RemoveItem((ushort)itemToRemove.Id);
         }
     }
 }
